Nisitha

+++ frontend/src/Components/Admin/SupplierTable.jsx
@@ -1,5 +1,8 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
+import { FaDeleteLeft } from "react-icons/fa6";
+import { MdEdit } from "react-icons/md";
+import LoadingAnimation from "../Login/LoadingAnimation";
 
 const SupplierComponent = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -21,6 +24,8 @@ const SupplierComponent = () => {
     mobile: "",
   });
   const [categories, setCategories] = useState([]);
+  const [loading, setLoading] = useState(true);
+
 
   // Fetch all suppliers and categories from the API
   useEffect(() => {
@@ -37,8 +42,15 @@ const SupplierComponent = () => {
           "http://localhost:4500/portaldev/readcategories"
         );
         setCategories(categoryResponse.data.data);
+
+
+        // Simulate minimum loading time
+        await new Promise((resolve) => setTimeout(resolve, 1000)); // Wait for at least 1000ms
+
       } catch (error) {
         console.error("Error fetching suppliers or categories:", error);
+      } finally {
+        setLoading(false); // Stop the loading animation after both conditions are met
       }
     };
 
@@ -87,7 +99,7 @@ const SupplierComponent = () => {
       ...prevFilterData,
       [name]: value,
     }));
-    
+
     // Apply filter immediately when input changes
     handleFilterSubmit({ ...filterData, [name]: value });
   };
@@ -136,21 +148,23 @@ const SupplierComponent = () => {
     }
   };
 
-  const handleDeleteSupplier = async (SRno,id) => {
-    try {
-      const response = await axios.delete(
-        `http://localhost:4500/portaldev/deleteSupplier/${id}`
-      );
-      if (response.status === 200) {
-        setSuppliers((prevSuppliers) =>
-          prevSuppliers.filter((supplier) => supplier.SRno !== SRno)
+  const handleDeleteSupplier = async (SRno, id) => {
+    if (window.confirm('Are you sure you want to delete this supplier?')) {
+      try {
+        const response = await axios.delete(
+          `http://localhost:4500/portaldev/deleteSupplier/${id}`
         );
-        setFilteredSuppliers((prevSuppliers) =>
-          prevSuppliers.filter((supplier) => supplier.SRno !== SRno)
-        );
+        if (response.status === 200) {
+          setSuppliers((prevSuppliers) =>
+            prevSuppliers.filter((supplier) => supplier.SRno !== SRno)
+          );
+          setFilteredSuppliers((prevSuppliers) =>
+            prevSuppliers.filter((supplier) => supplier.SRno !== SRno)
+          );
+        }
+      } catch (error) {
+        console.error("Error deleting supplier:", error.response ? error.response.data : error.message);
       }
-    } catch (error) {
-      console.error("Error deleting supplier:", error.response ? error.response.data : error.message);
     }
   };
 
@@ -171,10 +185,14 @@ const SupplierComponent = () => {
   };
 
   return (
-    <>
+    <>{loading ? (
+      <>
+        <LoadingAnimation />
+      </>
+    ) : (
       <div className="float-right w-full min-h-screen">
         <h2 className="flex justify-center text-black font-bold text-2xl mt-4">Supplier Table</h2>
-        <div className="mx-8 mt-4">
+        <div className="mx-8 my-5">
           <div>
             <div className="flex mb-4 space-x-2">
               <input
@@ -183,7 +201,7 @@ const SupplierComponent = () => {
                 placeholder="Filter by SR No"
                 value={filterData.SRno}
                 onChange={handleFilterChange}
-                className="flex-1 p-2 border border-gray-300 rounded focus:outline-none focus:border-green-500"
+                className="flex-1 p-2 border-2 border-gray-300 rounded focus:outline-none focus:border-2 focus:border-green-600"
               />
               <input
                 type="text"
@@ -191,7 +209,7 @@ const SupplierComponent = () => {
                 placeholder="Filter by Name"
                 value={filterData.name}
                 onChange={handleFilterChange}
-                className="flex-1 p-2 border border-gray-300 rounded focus:outline-none focus:border-green-500"
+                className="flex-1 p-2 border-2 border-gray-300 rounded focus:outline-none focus:border-2 focus:border-green-600"
               />
               <input
                 type="text"
@@ -199,7 +217,7 @@ const SupplierComponent = () => {
                 placeholder="Filter by category"
                 value={filterData.category}
                 onChange={handleFilterChange}
-                className="flex-1 p-2 border border-gray-300 rounded focus:outline-none focus:border-green-500"
+                className="flex-1 p-2 border-2 border-gray-300 rounded focus:outline-none focus:border-2 focus:border-green-600"
               />
               <input
                 type="text"
@@ -207,10 +225,10 @@ const SupplierComponent = () => {
                 placeholder="Filter by mobile"
                 value={filterData.mobile}
                 onChange={handleFilterChange}
-                className="flex-1 p-2 border border-gray-300 rounded focus:outline-none focus:border-green-500"
+                className="flex-1 p-2 border-2 border-gray-300 rounded focus:outline-none focus:border-2 focus:border-green-600"
               />
               <button
-                className="bg-green-500 text-white font-bold py-2 px-4 rounded-md hover:bg-green-600 focus:outline-none"
+                className="bg-green-800 hover:ring-2 ring-green-500 text-green-200 font-semibold px-5 py-2 rounded-lg duration-200"
                 onClick={() => handleOpenModal()}
               >
                 New
@@ -218,36 +236,30 @@ const SupplierComponent = () => {
             </div>
           </div>
 
-          <table className="min-w-full bg-white border border-gray-200">
+          <table className="min-w-full border border-collapse table-auto bg-gradient-to-r from-white via-gray-100 to-white rounded-xl overflow-hidden shadow-lg">
             <thead>
-              <tr className="bg-blue-500 text-white">
-                <th className="py-2 px-4 border">Supplier SR No</th>
-                <th className="py-2 px-4 border">Supplier Name</th>
-                <th className="py-2 px-4 border">Category</th>
-                <th className="py-2 px-4 border">Mobile</th>
-                <th className="py-2 px-4 border">Action</th>
+              <tr className="bg-gradient-to-r from-slate-900 to-indigo-600 text-white text-sm tracking-wide">
+                <th className="py-4 px-4 font-bold uppercase border">Supplier SR No</th>
+                <th className="py-4 px-4 font-bold uppercase border">Supplier Name</th>
+                <th className="py-4 px-4 font-bold uppercase border">Category</th>
+                <th className="py-4 px-4 font-bold uppercase border">Mobile</th>
+                <th className="py-4 px-4 font-bold uppercase border">Action</th>
               </tr>
             </thead>
             <tbody>
               {filteredSuppliers.length > 0 ? (
                 filteredSuppliers.map((supplier) => (
-                  <tr key={supplier.SRno}>
-                    <td className="py-2 px-4 border">{supplier.SRno}</td>
-                    <td className="py-2 px-4 border">{supplier.name}</td>
-                    <td className="py-2 px-4 border">{supplier.category}</td>
-                    <td className="py-2 px-4 border">{supplier.mobile}</td>
-                    <td className="py-2 px-4 border text-center">
-                      <button
-                        className="bg-yellow-500 text-white py-1 px-2 rounded hover:bg-yellow-600"
-                        onClick={() => handleOpenModal(supplier)}
-                      >
-                        Update
+                  <tr key={supplier.SRno} className="hover:bg-gray-200 transition-all duration-300 ease-in-out">
+                    <td className="py-4 px-4 font-semibold border">{supplier.SRno}</td>
+                    <td className="py-4 px-4 font-semibold border">{supplier.name}</td>
+                    <td className="py-4 px-4 font-semibold border">{supplier.category}</td>
+                    <td className="py-4 px-4 font-semibold border">{supplier.mobile}</td>
+                    <td className="py-4 px-2 font-semibold border text-center space-x-6">
+                      <button onClick={() => handleOpenModal(supplier)} >
+                        <MdEdit size={27} className="text-indigo-600 hover:scale-110" />
                       </button>
-                      <button
-                        className="bg-red-500 text-white py-1 px-2 rounded hover:bg-red-600 ml-2"
-                        onClick={() => handleDeleteSupplier(supplier.SRno, supplier._id)}
-                      >
-                        Delete
+                      <button onClick={() => handleDeleteSupplier(supplier.SRno, supplier._id)} >
+                        <FaDeleteLeft size={27} className="text-red-600 hover:scale-110" />
                       </button>
                     </td>
                   </tr>
@@ -284,7 +296,7 @@ const SupplierComponent = () => {
                       type="text"
                       name="name"
                       value={formData.name}
-                      onChange={handleInputChange}                      className="w-full p-2 border border-gray-300 rounded"
+                      onChange={handleInputChange} className="w-full p-2 border border-gray-300 rounded"
                     />
                   </div>
                   <div className="mb-4">
@@ -314,18 +326,18 @@ const SupplierComponent = () => {
                       onChange={handleInputChange}
                       className="w-full p-2 border border-gray-300 rounded"
                     ></textarea>
-                  </div> 
+                  </div>
                   <div className="flex justify-end space-x-4">
                     <button
                       type="submit"
-                      className="bg-blue-500 text-white py-2 px-4 rounded hover:bg-blue-600"
+                      className="text-blue-200 font-semibold px-5 py-2 rounded-lg bg-blue-800 hover:ring-2 ring-blue-500 duration-200"
                     >
                       {isEditMode ? "Update" : "Add"}
                     </button>
                     <button
                       type="button"
                       onClick={handleCloseModal}
-                      className="bg-gray-500 text-white py-2 px-4 rounded hover:bg-gray-600"
+                      className="text-gray-200 font-semibold px-5 py-2 rounded-lg bg-gray-800 hover:ring-2 ring-gray-500 duration-200"
                     >
                       Cancel
                     </button>
@@ -336,6 +348,7 @@ const SupplierComponent = () => {
           )}
         </div>
       </div>
+    )}
     </>
   );
 };
